ip statusbar

+++ install.sh
@@ -7,15 +7,18 @@ usage() {
 	echo "$EXEC_PATH/$EXEC_NAME -y: Auto Accept Dialogs"
 }
 
-change_shell_prompt() {
-	printf "Change Default Shell? [y/n] "
+prompt() {
+	$AUTO_ACCEPT && return 0
+	printf "$1? [y/n] "
 	read -r REPLY
 	REPLY=$(echo "$REPLY" | tr '[:upper:]' '[:lower:]')
 	[ "$REPLY" = "y" ] && return 0
 	[ "$REPLY" = "n" ] && return 1
 	echo "Bad Answer, Try Again!"
-	change_shell_prompt # Retry prompt if bad answer
+	prompt "$1" # Retry prompt if bad answer
 }
+
+alias reload_i3="(command -v pgrep i3-msg && pgrep '^i3$' && i3-msg reload)"
 
 export AUTO_ACCEPT=false
 
@@ -27,24 +30,19 @@ while getopts ":hy" o; do
 	esac
 done
 
-EXECS=$(find src -type f -exec awk '/^#!.*sh$/{print FILENAME}' {} \;)
-for file in $EXECS
-do
-	chmod +x "$file" || exit $?
-done
-cd src || exit $?
-	cp --preserve=all -r . "$HOME/"
-cd - > /dev/null || exit $?
+find src -type f -exec awk '/^#!.*sh$/{system("chmod +rwx " FILENAME)}' {} \;
+if prompt "Overwrite Config Files"; then
+	cd src || exit $?
+		cp --preserve=all -r . "$HOME/"
+	cd - > /dev/null || exit $?
+fi
 
-if [ -r "/etc/passwd" ]
-then
+if [ -r "/etc/passwd" ]; then
 	ZSH_PATH=$(command -v zsh || exit $?)
 	DEFAULT_SHELL=$(awk -F: "/^$USER:.*:$UID:$GID/{print \$7}" /etc/passwd)
 
-	if [ "$DEFAULT_SHELL" != "$ZSH_PATH" ]
-	then
-		if $AUTO_ACCEPT || change_shell_prompt
-		then
+	if [ "$DEFAULT_SHELL" != "$ZSH_PATH" ]; then
+		if prompt "Change Default Shell"; then
 			echo "Password Required to change shell"
 			chsh -s "$ZSH_PATH"
 		fi
@@ -55,7 +53,7 @@ else
 fi
 
 # Reload i3 config if running
-if ! (pgrep "^i3$" && i3-msg reload) > /dev/null
-then
-	echo "Error Reloading i3wm"
-fi
+reload_i3 > /dev/null
+
+systemctl --user daemon-reload
+systemctl --user enable --now pulseaudio
